+++ src/index.ts
@@ -1,9 +1,10 @@
 import 'isomorphic-fetch';
 import * as util from './util';
+import { formatIconUrlWithProxy } from './assets/assetsUtil';
 
 export * from './assets/AccountTrackerController';
 export * from './user/AddressBookController';
-export * from './approval/ApprovalController';
+export * from './approval';
 export * from './assets/AssetsContractController';
 export * from './BaseController';
 export {
@@ -39,4 +40,4 @@ export * from './permissions';
 export * from './subject-metadata';
 export * from './ratelimit/RateLimitController';
 export * from './notification/NotificationController';
-export { util };
+export { util, formatIconUrlWithProxy };