+++ CHANGELOG.md
@@ -6,6 +6,37 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [31.1.0]
+### Added
+- Add new error type - `ApprovalRequestNotFoundError` ([#909](https://github.com/MetaMask/controllers/pull/909))
+
+### Changed
+- Replace `ethers` with submodules (`@ethersproject/**`) ([#907](https://github.com/MetaMask/controllers/pull/907))
+
+## [31.0.0]
+### Added
+- Add support for caveats on endowment permissions ([#878](https://github.com/MetaMask/controllers/pull/878))
+
+### Changed
+- **BREAKING:** Bump to Node 14 ([#893](https://github.com/MetaMask/controllers/pull/893))
+- **BREAKING:** PhishingController is now async ([#897](https://github.com/MetaMask/controllers/pull/897))
+  - `PhishingController.test` now returns a Promise
+  - `PhishingController.setRefreshInterval` should now be used to change the polling interval
+- Changed TokenController._createEthersContract to no longer be async ([#895](https://github.com/MetaMask/controllers/pull/895))
+
+## [30.3.0]
+### Added
+- Add new functionality to the CollectiblesController to verify the transaction state of a collectible ([#762](https://github.com/MetaMask/controllers/pull/762)).
+  - Add new method to the CollectiblesController to search a Collectible by smart contract address and token ID (`findCollectibleByAddressAndTokenId`).
+  - Add new method to the CollectiblesController to update the state of a Collectible (`updateCollectible`).
+  - Add new method to the CollectiblesController to reset the transaction status of a Collectible (`resetCollectibleTransactionStatusByTransactionId`).
+- Add new method `validatePassword` to KeyringController to validate the password ([#884](https://github.com/MetaMask/controllers/pull/884))
+- Add `formatIconUrlWithProxy` function for formatting iconUrl Proxy for tokens from Swaps API ([#889](https://github.com/MetaMask/controllers/pull/889))
+
+## [30.2.0]
+### Changed
+- Updated the Testnet Ticker string format to be prefixed with the network name (e.g. <network name>ETH) ([#868](https://github.com/MetaMask/controllers/pull/868))
+
 ## [30.1.0]
 ### Added
 - Allow clients to avoid third party tokenlist polling ([#861](https://github.com/MetaMask/controllers/pull/861)):
@@ -634,7 +665,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Remove shapeshift controller (#209)
 
-[Unreleased]: https://github.com/MetaMask/controllers/compare/v30.1.0...HEAD
+[Unreleased]: https://github.com/MetaMask/controllers/compare/v31.1.0...HEAD
+[31.1.0]: https://github.com/MetaMask/controllers/compare/v31.0.0...v31.1.0
+[31.0.0]: https://github.com/MetaMask/controllers/compare/v30.3.0...v31.0.0
+[30.3.0]: https://github.com/MetaMask/controllers/compare/v30.2.0...v30.3.0
+[30.2.0]: https://github.com/MetaMask/controllers/compare/v30.1.0...v30.2.0
 [30.1.0]: https://github.com/MetaMask/controllers/compare/v30.0.2...v30.1.0
 [30.0.2]: https://github.com/MetaMask/controllers/compare/v30.0.1...v30.0.2
 [30.0.1]: https://github.com/MetaMask/controllers/compare/v30.0.0...v30.0.1